docker/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
